prepare search

+++ utils/configs.py
@@ -97,6 +97,6 @@ class Arguments:
     max_n_samples: int = field(default=10000000000)
     
     # bon
-    bon_rate: float = field(default=1)
+    bon_rate: int = field(default=1)
     bon_guidance: str = field(default=None)
